+++ js/main.js
@@ -1,5 +1,5 @@
 /* eslint-disable no-prototype-builtins */
-/* global getBlockClick  myBlockNumbering:writable objectlength config initVersion loadSettings settings getRandomInt number_format levelNamesEncoded _TEMP_SYMBOL hexToHsb Cookies*/
+/* global getBlockClick  myBlockNumbering:writable objectlength config initVersion loadSettings settings getRandomInt number_format levelNamesEncoded _TEMP_SYMBOL hexToHsb Cookies switchDevice*/
 /* global sessionValid MobileDetect moment getBlock buttons handleObjectBlock getGraphs iconORimage getBlockData titleAndValueSwitch showUpdateInformation getStateBlock addThermostatFunctions*/
 /* global loadWeatherFull loadWeather Swiper ion */
 
@@ -2188,9 +2188,9 @@ function getDimmerBlock(device, idx, buttonimg) {
     var html = '';
     var classExtension = isProtected(idx) ? ' icon' : ' icon iconslider'; //no pointer in case of protected device
     if (device['Status'] === 'Off')
-        html += iconORimage(idx, 'far fa-lightbulb', buttonimg, getIconStatusClass(device['Status']) + classExtension, '', 2, 'data-light="' + device['idx'] + '" onclick="switchDevice(this,\'toggle\', false );"');
+        html += iconORimage(idx, 'far fa-lightbulb', buttonimg, getIconStatusClass(device['Status']) + classExtension, '', 2, 'data-light="' + device['idx'] + '" ');
     else
-        html += iconORimage(idx, 'fas fa-lightbulb', buttonimg, getIconStatusClass(device['Status']) + classExtension, '', 2, 'data-light="' + device['idx'] + '" onclick="switchDevice(this,\'toggle\', false);"');
+        html += iconORimage(idx, 'fas fa-lightbulb', buttonimg, getIconStatusClass(device['Status']) + classExtension, '', 2, 'data-light="' + device['idx'] + '" ');
     html += '<div class="col-xs-10 swiper-no-swiping col-data">';
     html += '<strong class="title">' + device['Name'];
     if (typeof (blocks[idx]) == 'undefined' || typeof (blocks[idx]['hide_data']) == 'undefined' || blocks[idx]['hide_data'] == false) {
@@ -2215,8 +2215,16 @@ function getDimmerBlock(device, idx, buttonimg) {
     }
 
     $('div.block_' + idx).html(html);
+
+    var dimmerClickHandler = function() {
+        if (!sliding) switchDevice('.block_'+idx,'toggle', false )
+    }
+
+    $('div.block_' + idx).off('click');
+
     if (!isProtected(idx)) {
         $('div.block_' + idx).addClass('hover');
+        $('div.block_' + idx).on('click', dimmerClickHandler);
     }
 
     if (isRGBDeviceAndEnabled(device)) {
@@ -2390,9 +2398,15 @@ function addSlider(idx, sliderValues) {
             slideDeviceExt($(this).data('light'), ui.value, 2);
         },
         stop: function () {
-            sliding = false;
+            setTimeout(function() {
+                sliding = false;
+            }, 100); //prevent clickhandler of container by setting sliding to false only after 100ms
         }
     });
+    $(".slider" + idx).on('click', function(ev) {
+        ev.stopPropagation();
+    })
+
 }
 
 function isRGBDeviceAndEnabled(device) {